Re-triggering animation with key prop
The animation is not played when the component re-render but when it is re-mounted. That is where a useful usage of the key prop comes into play.
By providing the React element (i.e.: the same component used as an element) with a dynamic key we can trick React into thinking that the component has been unmounted, then triggering remount, thus re-triggering of the animation.